temp/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
Werkzeug==2.3.7
Click==8.1.7
